modules/advance_demo/requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
pydantic==1.9.1
pydotenv==0.0.7
PyMySQL==1.0.2
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.18.1
